fixed incorrect variable sizes

+++ dio/controllers/ite8783.cpp
@@ -56,7 +56,7 @@ Ite8783::~Ite8783()
 void Ite8783::initPin(PinInfo info)
 {
 	setSioLdn(kGpioLdn);
-	uint16_t reg = kPolarityBar + info.offset;
+	uint8_t reg = kPolarityBar + info.offset;
 	if (reg <= kPolarityMax)
 		writeSioRegister(reg, readSioRegister(reg) & ~info.bitmask);	//Set polarity to non-inverting
 
@@ -73,7 +73,7 @@ void Ite8783::initPin(PinInfo info)
 PinMode Ite8783::getPinMode(PinInfo info)
 {
 	setSioLdn(kGpioLdn);
-	uint16_t reg = kOutputEnableBar + info.offset;
+	uint8_t reg = kOutputEnableBar + info.offset;
 	uint8_t data = readSioRegister(reg);
 	if ((data & info.bitmask) == info.bitmask) 
 		return ModeOutput;
@@ -90,7 +90,7 @@ void Ite8783::setPinMode(PinInfo info, PinMode mode)
 		throw DioControllerError("Output mode not supported on pin");
 
 	setSioLdn(kGpioLdn);
-	uint16_t reg = kOutputEnableBar + info.offset;
+	uint8_t reg = kOutputEnableBar + info.offset;
 	uint8_t data = readSioRegister(reg);
 	if (mode == ModeInput) data &= ~info.bitmask;
 	else if (mode == ModeOutput) data |= info.bitmask;
